Get rid of unnecessary casts;

+++ print-bgp.c
@@ -1679,7 +1679,7 @@ bgp_attr_print(netdissect_options *ndo,
                         case (AFNUM_INET<<8 | SAFNUM_RT_ROUTING_INFO):
                         case (AFNUM_INET<<8 | SAFNUM_MULTICAST_VPN):
                         case (AFNUM_INET<<8 | SAFNUM_MDT):
-			    if (tlen < (int)sizeof(struct in_addr)) {
+			    if (tlen < sizeof(struct in_addr)) {
                                 ND_PRINT((ndo, "invalid len"));
                                 tlen = 0;
                             } else {
@@ -1692,7 +1692,7 @@ bgp_attr_print(netdissect_options *ndo,
                         case (AFNUM_INET<<8 | SAFNUM_VPNUNICAST):
                         case (AFNUM_INET<<8 | SAFNUM_VPNMULTICAST):
                         case (AFNUM_INET<<8 | SAFNUM_VPNUNIMULTICAST):
-                            if (tlen < (int)(sizeof(struct in_addr)+BGP_VPN_RD_LEN)) {
+                            if (tlen < sizeof(struct in_addr)+BGP_VPN_RD_LEN) {
                                 ND_PRINT((ndo, "invalid len"));
                                 tlen = 0;
                             } else {
@@ -1708,7 +1708,7 @@ bgp_attr_print(netdissect_options *ndo,
                         case (AFNUM_INET6<<8 | SAFNUM_MULTICAST):
                         case (AFNUM_INET6<<8 | SAFNUM_UNIMULTICAST):
                         case (AFNUM_INET6<<8 | SAFNUM_LABUNICAST):
-                            if (tlen < (int)sizeof(struct in6_addr)) {
+                            if (tlen < sizeof(struct in6_addr)) {
                                 ND_PRINT((ndo, "invalid len"));
                                 tlen = 0;
                             } else {
@@ -1721,7 +1721,7 @@ bgp_attr_print(netdissect_options *ndo,
                         case (AFNUM_INET6<<8 | SAFNUM_VPNUNICAST):
                         case (AFNUM_INET6<<8 | SAFNUM_VPNMULTICAST):
                         case (AFNUM_INET6<<8 | SAFNUM_VPNUNIMULTICAST):
-                            if (tlen < (int)(sizeof(struct in6_addr)+BGP_VPN_RD_LEN)) {
+                            if (tlen < sizeof(struct in6_addr)+BGP_VPN_RD_LEN) {
                                 ND_PRINT((ndo, "invalid len"));
                                 tlen = 0;
                             } else {
@@ -1737,7 +1737,7 @@ bgp_attr_print(netdissect_options *ndo,
                         case (AFNUM_L2VPN<<8 | SAFNUM_VPNUNICAST):
                         case (AFNUM_L2VPN<<8 | SAFNUM_VPNMULTICAST):
                         case (AFNUM_L2VPN<<8 | SAFNUM_VPNUNIMULTICAST):
-                            if (tlen < (int)sizeof(struct in_addr)) {
+                            if (tlen < sizeof(struct in_addr)) {
                                 ND_PRINT((ndo, "invalid len"));
                                 tlen = 0;
                             } else {
